fetching refactor + lxns data
