+++ src/app/api/admin/image/route.ts
@@ -7,16 +7,23 @@ import { nanoid } from "nanoid";
 import { NextRequest, NextResponse } from "next/server";
 
 const MAIMAI_COVER_PATTERN = /^https?:\/\/maimaidx(?:-eng)?\.com\/maimai-mobile\/img\/Music\/(.+)$/;
+// Lxns CN jacket: https://assets2.lxns.net/maimai/jacket/{id}.png — namespace under
+// `lxns_{id}` to avoid collisions with the JP/INTL md5-style filenames.
+const LXNS_COVER_PATTERN = /^https?:\/\/assets2?\.lxns\.net\/maimai\/jacket\/(\d+)\.png$/;
 
 const STATIC_ASSETS: { url: string; basename: string }[] = [
   { url: "https://maimaidx.jp/maimai-mobile/img/music_dx.png", basename: "music_dx" },
   { url: "https://maimaidx.jp/maimai-mobile/img/music_standard.png", basename: "music_standard" },
 ];
 
+// Returns a stable "filename" key for a cover URL (used for de-dup and R2 lookup).
+// Returns null for URLs we don't know how to cache.
 function extractFilename(url: string): string | null {
-  const match = url.match(MAIMAI_COVER_PATTERN);
-  if (!match) return null;
-  return match[1];
+  const maimai = url.match(MAIMAI_COVER_PATTERN);
+  if (maimai) return maimai[1];
+  const lxns = url.match(LXNS_COVER_PATTERN);
+  if (lxns) return `lxns_${lxns[1]}.png`;
+  return null;
 }
 
 function isJpDomain(url: string): boolean {
