+++ src/components/token-dialog-cn.tsx
@@ -9,7 +9,7 @@ import {
   ResponsiveDialogTitle,
 } from "@/components/ui/dialog-friendly";
 import { cn } from "@/lib/utils";
-import { ChevronRight, Fish, Key, Snowflake } from "lucide-react";
+import { ChevronRight, Fish, Key, Smartphone, Snowflake, Wifi } from "lucide-react";
 import { useTranslations } from "next-intl";
 
 interface TokenDialogCnProps {
@@ -23,14 +23,30 @@ const FETCH_OPTIONS = [
     name: "水鱼查分器",
     description: "使用您的水鱼查分器登入用户名 或 已绑定的 QQ 帐户 ID 取得资料。",
     recommended: false,
+    supported: false,
     icon: <Fish className="h-5 w-5 text-primary" />
   },
   {
     name: "落雪咖啡屋查分器",
     description: "以落雪咖啡屋帐号登入授权取得资料。",
     recommended: false,
+    supported: true,
     icon: <Snowflake className="h-5 w-5 text-primary" />
   },
+  {
+    name: "HTTP 代理",
+    description: "透过 HTTP 代理拦截手机微信中的舞萌 DX 小程序请求以取得资料。",
+    recommended: false,
+    supported: false,
+    icon: <Wifi className="h-5 w-5 text-primary" />
+  },
+  {
+    name: "Android 应用程式",
+    description: "在 Android 装置上安装本应用程式以自动取得资料。",
+    recommended: false,
+    supported: false,
+    icon: <Smartphone className="h-5 w-5 text-primary" />
+  }
 ] as const
 
 export function TokenDialogCn({
@@ -72,7 +88,8 @@ export function TokenDialogCn({
             {...FETCH_OPTIONS.map(option => (
               <button
                 onClick={() => { }}
-                className="w-full p-4 border-2 rounded-lg hover:border-primary hover:bg-accent/50 transition-all text-left group"
+                disabled={!option.supported}
+                className="w-full p-4 border-2 rounded-lg hover:border-primary hover:bg-accent/50 transition-all text-left group disabled:opacity-50 disabled:cursor-not-allowed disabled:hover:border-border disabled:hover:bg-transparent"
               >
                 <div className="flex items-start space-x-3">
                   <div className="mt-1 p-2 rounded-md bg-primary/10">
@@ -83,6 +100,9 @@ export function TokenDialogCn({
                       <span className="font-semibold text-base">{option.name}</span>
                       {option.recommended && (<Badge variant="default">
                         {t('tokenDialog.recommended')}
+                      </Badge>)}
+                      {!option.supported && (<Badge variant="secondary">
+                        即将推出
                       </Badge>)}
                     </div>
                     <p className="text-sm text-muted-foreground leading-relaxed">
