+++ src/lib/maimai/index.ts
@@ -1,0 +1,10 @@
+export { fetchMaimaiData } from "./orchestrator";
+export { upsertScoreData } from "./songs/persist";
+export type {
+  AlbumData,
+  EventAreaData,
+  EventData,
+  PlayerData,
+  RecentSongData,
+  ScoreData,
+} from "./types";
